modbus-rtu/Sanpropusk/requirements.txt
crcmod==1.7
future==1.0.0
iso8601==2.1.0
pymodbus==3.7.4
pyserial==3.5
PyYAML==6.0.2
requests
python-telegram-bot==20.7
